requirements.txt
customtkinter==5.2.2
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
Pillow==11.0.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
scikit_learn==1.6.0
tensorflow==2.17.0
